Docs: Separate Element definitions and utilities
