+++ packages/docs/elements/backgrounds/paper-texture/index.mdx
@@ -4,8 +4,8 @@ description: A white animated paper texture background.
 ---
 
 import {ElementPage} from '@site/src/components/Elements/ElementPage';
-import {paperTextureDefinition} from '@site/src/components/Elements/element-registry';
+import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
 
 # Paper Texture
 
-<ElementPage definition={paperTextureDefinition} sourceFile="./paper-texture.tsx" />
+<ElementPage definition={elementDefinitions['backgrounds/paper-texture']} sourceFile="./paper-texture.tsx" />
