+++ packages/docs/elements/backgrounds/rotating-starburst/index.mdx
@@ -4,8 +4,8 @@ description: A Solid background with a slowly rotating starburst effect.
 ---
 
 import {ElementPage} from '@site/src/components/Elements/ElementPage';
-import {rotatingStarburstDefinition} from '@site/src/components/Elements/element-registry';
+import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
 
 # Rotating Starburst
 
-<ElementPage definition={rotatingStarburstDefinition} sourceFile="./rotating-starburst.tsx" />
+<ElementPage definition={elementDefinitions['backgrounds/rotating-starburst']} sourceFile="./rotating-starburst.tsx" />
