+++ packages/docs/elements/overlays/lower-third/index.mdx
@@ -4,8 +4,8 @@ description: A clean animated lower third for names and titles.
 ---
 
 import {ElementPage} from '@site/src/components/Elements/ElementPage';
-import {lowerThirdDefinition} from '@site/src/components/Elements/element-registry';
+import {elementDefinitions} from '@site/src/components/Elements/element-definitions';
 
 # Lower Third
 
-<ElementPage definition={lowerThirdDefinition} sourceFile="./lower-third.tsx" />
+<ElementPage definition={elementDefinitions['overlays/lower-third']} sourceFile="./lower-third.tsx" />
