+++ packages/docs/elements/submit-an-element.mdx
@@ -33,7 +33,7 @@ Add:
 - An Element folder in `packages/docs/elements/<category>/<slug>/`
 - An `index.mdx` page
 - One self-contained TSX source file exporting a React component
-- An entry in `packages/docs/src/components/Elements/element-registry.ts`
+- An entry in `packages/docs/src/components/Elements/element-definitions.ts`
 
 For example:
 
@@ -54,14 +54,18 @@ The source file should work when dragged or copied into a project.
 
 ## 4. Define size
 
-To give an Element a fixed size, set `elementWidth` and `elementHeight` in its registry entry. Use `null` for both values if the Element should inherit the full composition dimensions.
+To give an Element a fixed size, set `elementWidth` and `elementHeight` in its definition entry. Use `null` for both values if the Element should inherit the full composition dimensions.
 
-```tsx title="element-registry.ts"
-export const myElementDefinition: ElementDefinition = {
-  // Other metadata...
-  elementWidth: 900,
-  elementHeight: 260,
-  previewPadding: 120,
+```tsx title="element-definitions.ts"
+export const elementDefinitions = {
+  // The object key must match the slug.
+  'overlays/my-element': {
+    // Other metadata...
+    slug: 'overlays/my-element',
+    elementWidth: 900,
+    elementHeight: 260,
+    previewPadding: 120,
+  },
 };
 ```
 
@@ -77,7 +81,7 @@ export const MyElement = () => {
 
 Size the Element to the smallest useful bounding box.
 
-Use `previewPadding` in the registry entry if the docs preview needs more breathing room without changing the Element's drag dimensions or adding source padding.
+Use `previewPadding` in the definition entry if the docs preview needs more breathing room without changing the Element's drag dimensions or adding source padding.
 
 Do not put the wrapper [`<Sequence>`](/docs/sequence) into the Element source file.
 
@@ -183,7 +187,7 @@ export const Badge = () => {
 
 ## 7. Check the preview assets
 
-Set an intentional `posterFrame` in the registry entry. It should show the Element in a representative, visible state.
+Set an intentional `posterFrame` in the definition entry. It should show the Element in a representative, visible state.
 
 Render the PNG and MP4 locally:
 
@@ -198,7 +202,7 @@ Inspect both files under `.element-previews/`. The MP4 should cover the Element'
 
 When submitting an Element as a pull request, include:
 
-- The Element folder and registry entry
+- The Element folder and definition entry
 - What makes it useful
 - The rendered preview
 - Links to inspiration, if any
